backend/requirements.txt
videodb[capture]>=0.4.0
flask
pycloudflared
python-dotenv
requests
google-genai
